+++ doc/version_history.rst
@@ -9,8 +9,9 @@ Version History
 v4.2.0
 ------
 
-* Add `enums.LaserTracker`.
-* Add PROPAGATINGBURSTMODEWAITINGFORTRIGGER and PROPAGATINGBURSTMODETRIGGERED to LaserDetailedState
+* Add ``enums.LaserTracker``.
+* `enums.TunableLaser.LaserDetailedState`: add PROPAGATINGBURSTMODEWAITINGFORTRIGGER and PROPAGATINGBURSTMODETRIGGERED.
+* `enums.ATMonochromator.Grating`: fix enum values.
 
 v4.1.0
 ------